+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `L10N.load_tar` to support embedded localization resources.
 * Changed `ByteLength` to display unit symbols.
 * Ignore not found error on `cargo zng l10n` cleanup.